Moved apps related code to a different module.
